move entity and dto packages to model package
