+++ battle-service/src/main/java/by/pashkevich/mikhail/client/UserClient.java
@@ -1,6 +1,6 @@
 package by.pashkevich.mikhail.client;
 
-import by.pashkevich.mikhail.dto.UserDto;
+import by.pashkevich.mikhail.model.dto.UserDto;
 import feign.Param;
 import feign.RequestLine;
