+++ battle-service/src/main/java/by/pashkevich/mikhail/mapper/FieldMapper.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.mapper;
 
-import by.pashkevich.mikhail.dto.FieldDto;
-import by.pashkevich.mikhail.entity.Field;
+import by.pashkevich.mikhail.model.dto.FieldDto;
+import by.pashkevich.mikhail.model.entity.Field;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
