+++ battle-service/src/main/java/by/pashkevich/mikhail/mapper/MoveMapper.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.mapper;
 
-import by.pashkevich.mikhail.dto.MoveDto;
-import by.pashkevich.mikhail.entity.Step;
+import by.pashkevich.mikhail.model.dto.MoveDto;
+import by.pashkevich.mikhail.model.entity.Step;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
