+++ battle-service/src/main/java/by/pashkevich/mikhail/model/dto/ResponseExceptionDto.java
@@ -1,12 +1,12 @@
-package by.pashkevich.mikhail.dto;
+package by.pashkevich.mikhail.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class PlayerDto {
-    private String nickname;
+@Data
+public class ResponseExceptionDto {
+    private String message;
 }
