+++ battle-service/src/main/java/by/pashkevich/mikhail/model/dto/UserDto.java
@@ -1,4 +1,4 @@
-package by.pashkevich.mikhail.dto;
+package by.pashkevich.mikhail.model.dto;
 
 import lombok.Data;
