+++ battle-service/src/main/java/by/pashkevich/mikhail/model/entity/Step.java
@@ -1,12 +1,13 @@
-package by.pashkevich.mikhail.dto;
+package by.pashkevich.mikhail.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
-public class ResponseExceptionDto {
-    private String message;
+public class Step {
+    private Integer i;
+    private Integer j;
 }
