+++ battle-service/src/main/java/by/pashkevich/mikhail/model/entity/enums/BattleStatus.java
@@ -1,4 +1,4 @@
-package by.pashkevich.mikhail.enums;
+package by.pashkevich.mikhail.model.entity.enums;
 
 import java.util.List;
