+++ battle-service/src/main/java/by/pashkevich/mikhail/repository/FieldRepository.java
@@ -1,6 +1,6 @@
 package by.pashkevich.mikhail.repository;
 
-import by.pashkevich.mikhail.entity.Field;
+import by.pashkevich.mikhail.model.entity.Field;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
