+++ battle-service/src/main/java/by/pashkevich/mikhail/repository/ScheduleSettingRepository.java
@@ -1,6 +1,6 @@
 package by.pashkevich.mikhail.repository;
 
-import by.pashkevich.mikhail.entity.setting.ScheduleSetting;
+import by.pashkevich.mikhail.model.entity.setting.ScheduleSetting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
