+++ battle-service/src/main/java/by/pashkevich/mikhail/service/BattleService.java
@@ -1,9 +1,9 @@
 package by.pashkevich.mikhail.service;
 
-import by.pashkevich.mikhail.entity.Battle;
-import by.pashkevich.mikhail.entity.Step;
-import by.pashkevich.mikhail.entity.User;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Battle;
+import by.pashkevich.mikhail.model.entity.Step;
+import by.pashkevich.mikhail.model.entity.User;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 
 import java.util.List;
