+++ battle-service/src/main/java/by/pashkevich/mikhail/service/BattleVerifyService.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.service;
 
-import by.pashkevich.mikhail.entity.Battle;
-import by.pashkevich.mikhail.entity.User;
+import by.pashkevich.mikhail.model.entity.Battle;
+import by.pashkevich.mikhail.model.entity.User;
 
 public interface BattleVerifyService {
     void isAvailableOrThrow(Battle battle, User user);
