+++ battle-service/src/main/java/by/pashkevich/mikhail/service/FieldService.java
@@ -1,9 +1,9 @@
 package by.pashkevich.mikhail.service;
 
-import by.pashkevich.mikhail.entity.Field;
-import by.pashkevich.mikhail.entity.Step;
-import by.pashkevich.mikhail.enums.BattleStatus;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Field;
+import by.pashkevich.mikhail.model.entity.Step;
+import by.pashkevich.mikhail.model.entity.enums.BattleStatus;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 
 public interface FieldService {
     BattleStatus move(Field field, Step step, Value value);
