+++ battle-service/src/main/java/by/pashkevich/mikhail/service/FieldVerifyService.java
@@ -1,8 +1,8 @@
 package by.pashkevich.mikhail.service;
 
-import by.pashkevich.mikhail.entity.Field;
-import by.pashkevich.mikhail.entity.Step;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Field;
+import by.pashkevich.mikhail.model.entity.Step;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 
 public interface FieldVerifyService {
     boolean isWin(Field field, Value value);
