+++ battle-service/src/main/java/by/pashkevich/mikhail/service/util/BattleUtil.java
@@ -1,9 +1,9 @@
 package by.pashkevich.mikhail.service.util;
 
-import by.pashkevich.mikhail.entity.Battle;
-import by.pashkevich.mikhail.entity.User;
-import by.pashkevich.mikhail.enums.BattleStatus;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Battle;
+import by.pashkevich.mikhail.model.entity.User;
+import by.pashkevich.mikhail.model.entity.enums.BattleStatus;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 import by.pashkevich.mikhail.exception.IncorrectDataException;
 import lombok.experimental.UtilityClass;
