+++ battle-service/src/test/java/by/pashkevich/mikhail/service/data/IsWinArgumentsProvider.java
@@ -1,14 +1,14 @@
 package by.pashkevich.mikhail.service.data;
 
-import by.pashkevich.mikhail.entity.Field;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Field;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 
 import java.util.stream.Stream;
 
-import static by.pashkevich.mikhail.enums.Value.*;
+import static by.pashkevich.mikhail.model.entity.enums.Value.*;
 import static by.pashkevich.mikhail.service.CommonMethods.anyId;
 
 public class IsWinArgumentsProvider implements ArgumentsProvider {
