+++ battle-service/src/test/java/by/pashkevich/mikhail/service/impl/BattleVerifyServiceImplTest.java
@@ -1,8 +1,8 @@
 package by.pashkevich.mikhail.service.impl;
 
-import by.pashkevich.mikhail.entity.Battle;
-import by.pashkevich.mikhail.entity.User;
-import by.pashkevich.mikhail.enums.BattleStatus;
+import by.pashkevich.mikhail.model.entity.Battle;
+import by.pashkevich.mikhail.model.entity.User;
+import by.pashkevich.mikhail.model.entity.enums.BattleStatus;
 import by.pashkevich.mikhail.exception.AccessException;
 import by.pashkevich.mikhail.exception.BattleUnavailableException;
 import by.pashkevich.mikhail.exception.IncorrectDataException;
