+++ battle-service/src/test/java/by/pashkevich/mikhail/service/impl/FieldServiceImplTest.java
@@ -1,9 +1,9 @@
 package by.pashkevich.mikhail.service.impl;
 
-import by.pashkevich.mikhail.entity.Field;
-import by.pashkevich.mikhail.entity.Step;
-import by.pashkevich.mikhail.enums.BattleStatus;
-import by.pashkevich.mikhail.enums.Value;
+import by.pashkevich.mikhail.model.entity.Field;
+import by.pashkevich.mikhail.model.entity.Step;
+import by.pashkevich.mikhail.model.entity.enums.BattleStatus;
+import by.pashkevich.mikhail.model.entity.enums.Value;
 import by.pashkevich.mikhail.exception.IncorrectDataException;
 import by.pashkevich.mikhail.repository.FieldRepository;
 import by.pashkevich.mikhail.service.FieldVerifyService;
