+++ user-service/src/main/java/by/pashkevich/mikhail/controller/UserController.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.controller;
 
-import by.pashkevich.mikhail.dto.UserDto;
-import by.pashkevich.mikhail.entity.User;
+import by.pashkevich.mikhail.model.dto.UserDto;
+import by.pashkevich.mikhail.model.entity.User;
 import by.pashkevich.mikhail.mapper.UserMapper;
 import by.pashkevich.mikhail.service.UserService;
 import lombok.RequiredArgsConstructor;
