--- user-service/src/main/java/by/pashkevich/mikhail/enums/Rolename.java
@@ -1,5 +1,0 @@
-package by.pashkevich.mikhail.enums;
-
-public enum Rolename {
-    USER
-}
