+++ user-service/src/main/java/by/pashkevich/mikhail/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.mapper;
 
-import by.pashkevich.mikhail.dto.UserDto;
-import by.pashkevich.mikhail.entity.User;
+import by.pashkevich.mikhail.model.dto.UserDto;
+import by.pashkevich.mikhail.model.entity.User;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring", uses = RoleMapper.class)
