+++ user-service/src/main/java/by/pashkevich/mikhail/model/entity/Role.java
@@ -1,6 +1,6 @@
-package by.pashkevich.mikhail.entity;
+package by.pashkevich.mikhail.model.entity;
 
-import by.pashkevich.mikhail.enums.Rolename;
+import by.pashkevich.mikhail.model.entity.enums.Rolename;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
