+++ user-service/src/main/java/by/pashkevich/mikhail/model/entity/enums/Rolename.java
@@ -1,0 +1,5 @@
+package by.pashkevich.mikhail.model.entity.enums;
+
+public enum Rolename {
+    USER
+}
