+++ user-service/src/main/java/by/pashkevich/mikhail/repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package by.pashkevich.mikhail.repository;
 
-import by.pashkevich.mikhail.entity.Role;
-import by.pashkevich.mikhail.enums.Rolename;
+import by.pashkevich.mikhail.model.entity.Role;
+import by.pashkevich.mikhail.model.entity.enums.Rolename;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
