+++ user-service/src/main/java/by/pashkevich/mikhail/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package by.pashkevich.mikhail.repository;
 
-import by.pashkevich.mikhail.entity.User;
+import by.pashkevich.mikhail.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
