+++ user-service/src/main/java/by/pashkevich/mikhail/service/UserService.java
@@ -1,6 +1,6 @@
 package by.pashkevich.mikhail.service;
 
-import by.pashkevich.mikhail.entity.User;
+import by.pashkevich.mikhail.model.entity.User;
 
 public interface UserService {
     String getJwt(User user);
