+++ user-service/src/main/java/by/pashkevich/mikhail/service/impl/UserServiceImpl.java
@@ -1,8 +1,8 @@
 package by.pashkevich.mikhail.service.impl;
 
-import by.pashkevich.mikhail.entity.Role;
-import by.pashkevich.mikhail.entity.User;
-import by.pashkevich.mikhail.enums.Rolename;
+import by.pashkevich.mikhail.model.entity.Role;
+import by.pashkevich.mikhail.model.entity.User;
+import by.pashkevich.mikhail.model.entity.enums.Rolename;
 import by.pashkevich.mikhail.exception.DuplicateException;
 import by.pashkevich.mikhail.exception.NotFoundException;
 import by.pashkevich.mikhail.repository.RoleRepository;
